新增统计时长

+++ TrackRecorder/Classes/Category/UICollectionView+TrackRecord.m
@@ -31,7 +31,7 @@
     
     // 交换后的方法
     NSString *targetName = NSStringFromClass([[[TrackRecordManager sharedInstance] getViewControllerWithTarget:collectionView] class]);
-    NSDictionary *dic = [TrackRecordManager sharedInstance].configDic[@"CollectionViewEvents"][targetName][[NSString stringWithFormat:@"%ld",indexPath.section]][[NSString stringWithFormat:@"%ld",indexPath.row]];
+    NSDictionary *dic = [TrackRecordManager sharedInstance].configDic[@"CollectionViewEvents"][targetName][[NSString stringWithFormat:@"%ld",(long)indexPath.section]][[NSString stringWithFormat:@"%ld",(long)indexPath.row]];
     if (dic) {
         // 需要埋点
         if ([[TrackRecordManager sharedInstance].delegate respondsToSelector:@selector(TrackRecordClickWithDictionary:)]) {
